fix to uint8 t

+++ src/ros2/statemachine/bt_plugins/src/action/import_environment_led_color_action.cpp
@@ -15,13 +15,13 @@ namespace bt_plugins
 
   BT::NodeStatus ImportEnvironmentLEDColor::tick()
   {
-    auto getEnvAsInt = [](const char *varName, int defaultValue) -> int
+    auto getEnvAsInt = [](const char *varName, int defaultValue) -> uint8_t
     {
       const char *value = std::getenv(varName);
       if (value)
       {
         std::istringstream iss(value);
-        int intValue;
+        uint8_t intValue;
         if (iss >> intValue)
         {
           return intValue;
@@ -31,10 +31,10 @@ namespace bt_plugins
     };
 
     // Get the values from the environment variables
-    int red = getEnvAsInt("RED", 0);
-    int green = getEnvAsInt("GREEN", 0);
-    int blue = getEnvAsInt("BLUE", 0);
-    int brightness = getEnvAsInt("BRIGHTNESS", 0);
+    uint8_t red = getEnvAsInt("RED", 0);
+    uint8_t green = getEnvAsInt("GREEN", 0);
+    uint8_t blue = getEnvAsInt("BLUE", 0);
+    uint8_t brightness = getEnvAsInt("BRIGHTNESS", 0);
 
     setOutput("blue", blue);
     setOutput("red", red);
